+++ tests/Examples/ExampleException.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ExampleException class.
+ *
+ * @copyright   Copyright (c) 2021 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2021-05-06
+ *
+ */
+
+namespace Konekt\ExtLogger\Tests\Examples;
+
+use Exception;
+
+class ExampleException extends Exception
+{
+}